+++ src/iceberg/iceberg_parser/iceberg_parser.py
@@ -9,6 +9,15 @@ import argparse
 import sys
 import json
 
+from .seals_parser import SealsSubparser
+from .penguins_parser import PenguinsSubparser
+from .rivers_parser import RiversSubparser
+from .landcover_parser import LandcoverSubparser
+
+PARSERS = {'seals': SealsSubparser,
+           'penguins': PenguinsSubparser,
+           'rivers': RiversSubparser,
+           'landcover': LandcoverSubparser}
 
 class IcebergParser(object):
     """
@@ -63,89 +72,10 @@ class IcebergParser(object):
                                        help='The project ID to charge',
                                        type=str, default=None)
 
-            command_args = parser.add_subparsers(help='commands')
-            seals_parser = command_args.add_parser('seals')
-            penguins_parser = command_args.add_parser('penguins')
-            four_d_geoloc_parser = command_args.add_parser('4Dgeolocation')
-            rivers_parser = command_args.add_parser('rivers')
-            landcover_parser = command_args.add_parser('landcover')
-
-            seals_parser.description = 'These are the options for Seals type \
-                                       analysis.'
-            seals_parser.set_defaults(which='seals')
-            seals_parser.add_argument('--scale_bands', '-s',
-                                      help='The size of the scale bands')
-            seals_parser.add_argument('--model', '-m',
-                                      help='The size of the scale bands')
-            seals_parser.add_argument('--model_path', '-mp',
-                                      help='Path of a custom model')
-            seals_parser.add_argument('--model_arch', '-ma',
-                                      help='Model Architecture')
-            seals_parser.add_argument('--hyperparameters', '-hy',
-                                      help='Hyperparameter Set')
-
-            penguins_parser.description = 'These are the options for Penguins \
-                                           type analysis.'
-            penguins_parser.set_defaults(which='penguins')
-            penguins_parser.add_argument('--scale_bands', '-s',
-                                         help='The size of the scale bands')
-            penguins_parser.add_argument('--model', '-m',
-                                         help='The size of the scale bands')
-            penguins_parser.add_argument('--model_path', '-mp',
-                                         help='Path of a custom model')
-            penguins_parser.add_argument('--hyperparameters', '-hy',
-                                         help='Hyperparameter Set')
-            penguins_parser.add_argument('--shadow_mask', '-sm')
-
-            four_d_geoloc_parser.description = 'These are the options for \
-                                                4DGeolocaltion type analysis.'
-            four_d_geoloc_parser.set_defaults(which='4DGeolocation')
-            four_d_geoloc_parser.add_argument('--target_path', '-t',
-                                              help='Path to target images')
-            four_d_geoloc_parser.add_argument('--threshold', '-th',
-                                              help='Minimum, maximum number of \
-                                              match points')
-            four_d_geoloc_parser.add_argument('--pixel_accuracy', '-pa',
-                                              help='An accuracy threshold for \
-                                              output pixels when making a \
-                                              match')
-            four_d_geoloc_parser.add_argument('--source_image_window', '-siw',
-                                              help='Subset window of the \
-                                              source image to search within')
-            four_d_geoloc_parser.add_argument('--target_image_window', '-tiw',
-                                              help='Subset window of the \
-                                              target image to search within')
-            four_d_geoloc_parser.add_argument('--algorithm', '-a',
-                                              help='which keypoint search \
-                                              algorithm to use')
-
-            rivers_parser.description = 'These are the options for Rivers type \
-                                         analysis.'
-            rivers_parser.set_defaults(which='rivers')
-            rivers_parser.add_argument('--threshold', '-th',
-                                       help='Minimum confidence to accept')
-            rivers_parser.add_argument('--hyperparameters', '-hy')
-            rivers_parser.add_argument('--model', '-m',
-                                       help='The size of the scale bands')
-            rivers_parser.add_argument('--model_path', '-mp',
-                                       help='Path of a custom model')
-            rivers_parser.add_argument('--ndwi_path', '-np',
-                                       help='Path to Water mask')
-
-            landcover_parser.description = 'These are the options for \
-                                            Landcover type analysis.'
-            landcover_parser.set_defaults(which='landcover')
-            landcover_parser.add_argument('--spec_lib', '-sl',
-                                          help='Addition of new ground data to \
-                                          spectral library')
-            landcover_parser.add_argument('--roi_sel', '-rs',
-                                          help='Selection of regions of \
-                                          interest for atmospheric correction')
-            landcover_parser.add_argument('--atmcorr_model', '-am',
-                                          help='Selection of atmospheric model')
-            landcover_parser.add_argument('--landcover_lib', '-ll',
-                                          help='Access landcover masks')
-            landcover_parser.add_argument('--shadow_mask', '-sm')
+            command_parser = parser.add_subparsers(help='commands')
+            
+            for key, parser_impl in PARSERS.iteritems():
+                parser_impl(command_parser)
 
             tmp_args = parser.parse_args()
             tmp_args = vars(tmp_args)
